+++ paymentsheet/src/main/java/com/stripe/android/link/ui/LinkContent.kt
@@ -5,17 +5,18 @@ import androidx.compose.animation.core.Spring
 import androidx.compose.animation.core.spring
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.IntrinsicSize
-import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CornerSize
-import androidx.compose.material.ModalBottomSheetLayout
-import androidx.compose.material.ModalBottomSheetState
+import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
@@ -23,6 +24,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.stripe.android.common.ui.AnimatedContentHeight
+import com.stripe.android.common.ui.ElementsBottomSheetLayout
 import com.stripe.android.core.utils.FeatureFlags
 import com.stripe.android.link.LinkAccountUpdate
 import com.stripe.android.link.LinkAction
@@ -35,7 +37,6 @@ import com.stripe.android.link.linkViewModel
 import com.stripe.android.link.model.LinkAccount
 import com.stripe.android.link.theme.DefaultLinkTheme
 import com.stripe.android.link.theme.LinkTheme
-import com.stripe.android.link.theme.LinkThemeConfig.scrim
 import com.stripe.android.link.ui.paymentmenthod.PaymentMethodScreen
 import com.stripe.android.link.ui.paymentmenthod.PaymentMethodViewModel
 import com.stripe.android.link.ui.signup.SignUpScreen
@@ -46,6 +47,8 @@ import com.stripe.android.link.ui.verification.VerificationScreen
 import com.stripe.android.link.ui.verification.VerificationViewModel
 import com.stripe.android.link.ui.wallet.WalletScreen
 import com.stripe.android.link.ui.wallet.WalletViewModel
+import com.stripe.android.uicore.elements.bottomsheet.rememberStripeBottomSheetState
+import kotlinx.coroutines.launch
 
 @SuppressWarnings("LongMethod")
 @Composable
@@ -53,9 +56,6 @@ internal fun LinkContent(
     modifier: Modifier,
     navController: NavHostController,
     appBarState: LinkAppBarState,
-    sheetState: ModalBottomSheetState,
-    bottomSheetContent: BottomSheetContent?,
-    onUpdateSheetContent: (BottomSheetContent?) -> Unit,
     handleViewAction: (LinkAction) -> Unit,
     navigate: (route: LinkScreen, clearStack: Boolean) -> Unit,
     dismissWithResult: (LinkActivityResult) -> Unit,
@@ -66,75 +66,92 @@ internal fun LinkContent(
     changeEmail: () -> Unit,
     initialDestination: LinkScreen
 ) {
+    val coroutineScope = rememberCoroutineScope()
+    var bottomSheetContent by remember { mutableStateOf<BottomSheetContent?>(null) }
+    val bottomSheetState = rememberStripeBottomSheetState()
+
+    val onUpdateSheetContent: (BottomSheetContent?) -> Unit = { content ->
+        if (content != null) {
+            bottomSheetContent = content
+        } else {
+            coroutineScope.launch {
+                bottomSheetState.hide()
+            }
+        }
+    }
+
+    LaunchedEffect(bottomSheetContent) {
+        if (bottomSheetContent != null) {
+            bottomSheetState.show()
+        }
+    }
+
     DefaultLinkTheme {
         Surface(
             modifier = modifier,
             color = LinkTheme.colors.surfacePrimary,
         ) {
             val mbsLayoutModifier =
-                if (FeatureFlags.linkDynamicBottomSheet.isEnabled)
-                    Modifier.height(IntrinsicSize.Min)
-                else
+                if (FeatureFlags.linkDynamicBottomSheet.isEnabled) {
+                    Modifier.wrapContentHeight()
+                } else {
                     Modifier.fillMaxSize()
-            ModalBottomSheetLayout(
-                sheetContent = bottomSheetContent ?: {
-                    // Must have some content at startup or bottom sheet crashes when
-                    // calculating its initial size
-                    Box(Modifier.defaultMinSize(minHeight = 1.dp)) {}
-                },
-                modifier = mbsLayoutModifier,
-                sheetState = sheetState,
-                sheetShape = LinkTheme.shapes.large.copy(
-                    bottomStart = CornerSize(0.dp),
-                    bottomEnd = CornerSize(0.dp)
-                ),
-                sheetBackgroundColor = LinkTheme.colors.surfacePrimary,
-                scrimColor = LinkTheme.colors.scrim
+                }
+            Column(
+                modifier = mbsLayoutModifier
+                    .fillMaxWidth()
             ) {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                ) {
-                    BackHandler {
-                        if (bottomSheetContent != null) {
-                            onUpdateSheetContent(null)
-                        } else if (navController.popBackStack().not()) {
-                            handleViewAction(LinkAction.BackPressed)
-                        }
-                    }
-
-                    LinkAppBar(
-                        state = appBarState,
-                        onBackPressed = onBackPressed,
-                        showBottomSheetContent = onUpdateSheetContent,
-                        onLogoutClicked = {
-                            handleViewAction(LinkAction.LogoutClicked)
-                        }
-                    )
-
-                    AnimatedContentHeight(
-                        animationSpec = spring(stiffness = Spring.StiffnessMediumLow),
-                        isEnabled = FeatureFlags.linkDynamicBottomSheet.isEnabled,
-                    ) {
-                        Screens(
-                            initialDestination = initialDestination,
-                            navController = navController,
-                            goBack = goBack,
-                            moveToWeb = moveToWeb,
-                            navigateAndClearStack = { screen ->
-                                navigate(screen, true)
-                            },
-                            dismissWithResult = dismissWithResult,
-                            getLinkAccount = getLinkAccount,
-                            showBottomSheetContent = onUpdateSheetContent,
-                            changeEmail = changeEmail,
-                            hideBottomSheetContent = {
-                                onUpdateSheetContent(null)
-                            }
-                        )
+                BackHandler {
+                    if (bottomSheetContent != null) {
+                        onUpdateSheetContent(null)
+                    } else if (navController.popBackStack().not()) {
+                        handleViewAction(LinkAction.BackPressed)
                     }
                 }
+
+                LinkAppBar(
+                    state = appBarState,
+                    onBackPressed = onBackPressed,
+                    showBottomSheetContent = onUpdateSheetContent,
+                    onLogoutClicked = {
+                        handleViewAction(LinkAction.LogoutClicked)
+                    }
+                )
+
+                AnimatedContentHeight(
+                    animationSpec = spring(stiffness = Spring.StiffnessMediumLow),
+                    isEnabled = FeatureFlags.linkDynamicBottomSheet.isEnabled,
+                ) {
+                    Screens(
+                        initialDestination = initialDestination,
+                        navController = navController,
+                        goBack = goBack,
+                        moveToWeb = moveToWeb,
+                        navigateAndClearStack = { screen ->
+                            navigate(screen, true)
+                        },
+                        dismissWithResult = dismissWithResult,
+                        getLinkAccount = getLinkAccount,
+                        showBottomSheetContent = onUpdateSheetContent,
+                        changeEmail = changeEmail,
+                        hideBottomSheetContent = {
+                            onUpdateSheetContent(null)
+                        }
+                    )
+                }
             }
+        }
+
+        bottomSheetContent?.let { content ->
+            ElementsBottomSheetLayout(
+                state = bottomSheetState,
+                onDismissed = { bottomSheetContent = null },
+                content = {
+                    DefaultLinkTheme {
+                        Column { content() }
+                    }
+                },
+            )
         }
     }
 }
